+++ examples/pagination/src/App.js
@@ -96,8 +96,9 @@ function Table({ columns, data }) {
         </thead>
         <tbody {...getTableBodyProps()}>
           {page.map(
-            (row, i) =>
-              prepareRow(row) || (
+            (row, i) => {
+              prepareRow(row);
+              return (
                 <tr {...row.getRowProps()}>
                   {row.cells.map(cell => {
                     return (
@@ -105,7 +106,7 @@ function Table({ columns, data }) {
                     )
                   })}
                 </tr>
-              )
+              )}
           )}
         </tbody>
       </table>