+++ src/plugin-hooks/useGroupBy.js
@@ -1,5 +1,4 @@
-import { useMemo } from 'react'
-import PropTypes from 'prop-types'
+import React from 'react'
 
 import * as aggregations from '../aggregations'
 import { addActions, actions } from '../actions'
@@ -15,27 +14,6 @@ import {
 defaultState.groupBy = []
 
 addActions('toggleGroupBy')
-
-const propTypes = {
-  // General
-  columns: PropTypes.arrayOf(
-    PropTypes.shape({
-      aggregate: PropTypes.oneOfType([
-        PropTypes.func,
-        PropTypes.string,
-        PropTypes.arrayOf(
-          PropTypes.oneOfType([PropTypes.func, PropTypes.string])
-        ),
-      ]),
-      disableGrouping: PropTypes.bool,
-      Aggregated: PropTypes.any,
-    })
-  ),
-  groupByFn: PropTypes.func,
-  manualGrouping: PropTypes.bool,
-  disableGrouping: PropTypes.bool,
-  aggregations: PropTypes.object,
-}
 
 export const useGroupBy = hooks => {
   hooks.columnsBeforeHeaderGroups.push(columnsBeforeHeaderGroups)
@@ -66,18 +44,20 @@ function columnsBeforeHeaderGroups(flatColumns, { state: { groupBy } }) {
   ]
 }
 
-function useMain(instance) {
-  PropTypes.checkPropTypes(propTypes, instance, 'property', 'useGroupBy')
+const defaultUserAggregations = {}
 
+function useMain(instance) {
   const {
     debug,
     rows,
+    flatRows,
     flatColumns,
     flatHeaders,
     groupByFn = defaultGroupByFn,
     manualGroupBy,
-    disableGrouping,
-    aggregations: userAggregations = {},
+    defaultCanGroupBy,
+    disableGroupBy,
+    aggregations: userAggregations = defaultUserAggregations,
     hooks,
     plugins,
     state: { groupBy },
@@ -87,17 +67,22 @@ function useMain(instance) {
   ensurePluginOrder(plugins, [], 'useGroupBy', ['useSortBy', 'useExpanded'])
 
   flatColumns.forEach(column => {
-    const { id, accessor, disableGrouping: columnDisableGrouping } = column
+    const {
+      id,
+      accessor,
+      defaultGroupBy: defaultColumnGroupBy,
+      disableGroupBy: columnDisableGroupBy,
+    } = column
     column.isGrouped = groupBy.includes(id)
     column.groupedIndex = groupBy.indexOf(id)
 
     column.canGroupBy = accessor
       ? getFirstDefined(
-          columnDisableGrouping,
-          disableGrouping === true ? false : undefined,
+          columnDisableGroupBy === true ? false : undefined,
+          disableGroupBy === true ? false : undefined,
           true
         )
-      : false
+      : getFirstDefined(defaultColumnGroupBy, defaultCanGroupBy, false)
 
     if (column.canGroupBy) {
       column.toggleGroupBy = () => toggleGroupBy(column.id)
@@ -125,6 +110,10 @@ function useMain(instance) {
 
   hooks.getGroupByToggleProps = []
 
+  // use reference to avoid memory leak in #1608
+  const instanceRef = React.useRef()
+  instanceRef.current = instance
+
   flatHeaders.forEach(header => {
     const { canGroupBy } = header
     header.getGroupByToggleProps = props => {
@@ -141,7 +130,11 @@ function useMain(instance) {
           },
           title: 'Toggle GroupBy',
         },
-        applyPropHooks(instance.hooks.getGroupByToggleProps, header, instance),
+        applyPropHooks(
+          instanceRef.current.hooks.getGroupByToggleProps,
+          header,
+          instanceRef.current
+        ),
         props
       )
     }
@@ -160,9 +153,9 @@ function useMain(instance) {
     return row
   })
 
-  const groupedRows = useMemo(() => {
+  const [groupedRows, groupedFlatRows] = React.useMemo(() => {
     if (manualGroupBy || !groupBy.length) {
-      return rows
+      return [rows, flatRows]
     }
 
     if (process.env.NODE_ENV === 'development' && debug)
@@ -170,7 +163,7 @@ function useMain(instance) {
     // Find the columns that can or are aggregating
 
     // Uses each column to aggregate rows into a single value
-    const aggregateRowsToValues = (rows, isSourceRows) => {
+    const aggregateRowsToValues = (rows, isAggregated) => {
       const values = {}
 
       flatColumns.forEach(column => {
@@ -191,7 +184,7 @@ function useMain(instance) {
               `React Table: Complex aggregators must have 2 values, eg. aggregate: ['sum', 'count']. More info above...`
             )
           }
-          if (isSourceRows) {
+          if (isAggregated) {
             aggregator = aggregator[1]
           } else {
             aggregator = aggregator[0]
@@ -204,7 +197,7 @@ function useMain(instance) {
             : userAggregations[aggregator] || aggregations[aggregator]
 
         if (aggregateFn) {
-          values[column.id] = aggregateFn(columnValues, rows)
+          values[column.id] = aggregateFn(columnValues, rows, isAggregated)
         } else if (aggregator) {
           console.info({ column })
           throw new Error(
@@ -217,10 +210,16 @@ function useMain(instance) {
       return values
     }
 
+    let groupedFlatRows = []
+
     // Recursively group the data
     const groupRecursively = (rows, depth = 0, parentPath = []) => {
       // This is the last level, just return the rows
       if (depth >= groupBy.length) {
+        rows.forEach(row => {
+          row.path = [...parentPath, ...row.path]
+        })
+        groupedFlatRows = groupedFlatRows.concat(rows)
         return rows
       }
 
@@ -236,10 +235,7 @@ function useMain(instance) {
 
           subRows = groupRecursively(subRows, depth + 1, path)
 
-          const values = aggregateRowsToValues(
-            subRows,
-            depth + 1 >= groupBy.length
-          )
+          const values = aggregateRowsToValues(subRows, depth < groupBy.length)
 
           const row = {
             isAggregated: true,
@@ -252,6 +248,8 @@ function useMain(instance) {
             path,
           }
 
+          groupedFlatRows.push(row)
+
           return row
         }
       )
@@ -259,13 +257,16 @@ function useMain(instance) {
       return groupedRows
     }
 
+    const groupedRows = groupRecursively(rows)
+
     // Assign the new data
-    return groupRecursively(rows)
+    return [groupedRows, groupedFlatRows]
   }, [
     manualGroupBy,
     groupBy,
     debug,
     rows,
+    flatRows,
     flatColumns,
     userAggregations,
     groupByFn,
@@ -275,6 +276,7 @@ function useMain(instance) {
     ...instance,
     toggleGroupBy,
     rows: groupedRows,
+    flatRows: groupedFlatRows,
     preGroupedRows: rows,
   }
 }