+++ src/plugin-hooks/useResizeColumns.js
@@ -1,6 +1,4 @@
-import PropTypes from 'prop-types'
-
-//
+import React from 'react'
 
 import { defaultState } from '../hooks/useTable'
 import { defaultColumn, getFirstDefined } from '../utils'
@@ -12,8 +10,6 @@ defaultState.columnResizing = {
 
 defaultColumn.canResize = true
 
-const propTypes = {}
-
 export const useResizeColumns = hooks => {
   hooks.useBeforeDimensions.push(useBeforeDimensions)
 }
@@ -21,8 +17,6 @@ export const useResizeColumns = hooks => {
 useResizeColumns.pluginName = 'useResizeColumns'
 
 const useBeforeDimensions = instance => {
-  PropTypes.checkPropTypes(propTypes, instance, 'property', 'useResizeColumns')
-
   instance.hooks.getResizerProps = []
 
   const {
@@ -99,6 +93,10 @@ const useBeforeDimensions = instance => {
     }))
   }
 
+  // use reference to avoid memory leak in #1608
+  const instanceRef = React.useRef()
+  instanceRef.current = instance
+
   flatHeaders.forEach(header => {
     const canResize = getFirstDefined(
       header.disableResizing === true ? false : undefined,
@@ -120,7 +118,11 @@ const useBeforeDimensions = instance => {
             },
             draggable: false,
           },
-          applyPropHooks(instance.hooks.getResizerProps, header, instance),
+          applyPropHooks(
+            instanceRef.current.hooks.getResizerProps,
+            header,
+            instanceRef.current
+          ),
           userProps
         )
       }