+++ CHANGELOG.md
@@ -1,3 +1,47 @@
+## 7.0.0-beta.19
+
+- Added an `isAggregated` boolean parameter to the `aggregate` function signature
+
+## 7.0.0-beta.16
+
+- Removed service workers from examples
+- Fixed a memory leak when `instance` was referenced in function closures
+- Fixed an issue where the table would infinitely rerender due to incorrect effect dependencies
+- Fixed an issue where row grouping and row selection would not work properly together.
+
+## 7.0.0-beta.15
+
+- Fixed an issue where `defaultGetResetPageDeps` was using `data` instead of `rows`
+
+## 7.0.0-beta.14
+
+- Removed
+  - `disablePageResetOnDataChange` option. use the `getResetPageDeps` option now.
+- Added
+  - `getResetPageDeps` option
+  - `getResetFilterDeps` option
+  - `getResetSortByDeps` option
+  - `getResetGroupByDeps` option
+  - `getResetExpandedDeps` option
+
+## 7.0.0-beta.13
+
+- Added options
+  - `defaultCanSort`
+  - `defaultCanFilter`
+  - `defaultCanGroupBy`
+  - `column.defaultCanSort`
+  - `column.defaultCanFilter`
+  - `column.defaultCanGroupBy`
+- Renamed
+  - `disableGrouping` to `disableGroupBy`
+  - `disableSorting` to `disableSortBy`
+  - `disableGroupBy` to `disableGroupBy`
+  - `column.disableGrouping` to `column.disableGroupBy`
+  - `column.disableSorting` to `column.disableSortBy`
+  - `column.disableGroupBy` to `column.disableGroupBy`
+- Removed propType definitions. Since types are now being maintained, it makes little sense to also maintain these. Cooincidentally, this also saves some bundle size in some scenarios where they may not be removed properly by a developer's bundler.
+
 ## 7.0.0-beta.0
 
 - Massive changes to the entire project and library. Please consult the README and documentation for more information regarding these changes.